fix: delete bug fix

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -150,9 +150,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        Image image = imageService.getImage(imageId);
+        Integer imageUserId = image.getUser().getId();
+        Integer sessionUserId = ((User)session.getAttribute("loggeduser")).getId();
+        if(sessionUserId.equals(imageUserId)) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+
+            model.addAttribute("deleteError", true);
+            model.addAttribute("image", image);
+            return "images/image";
+        }
     }
